Split CI jobs into more steps for clearer output
The main benefit (and motivation for this change) is from splitting
the old "Run Tests" step into an "Install project and dependencies"
step, followed by a "Run Tests" step.

With the new "Run Tests" step, it's easy to see the pytest output
in the GitHub interface for a run. Expanding that step shows all,
and only, the actual test output.

This allows the CI output itself to be as readily usable as the
output of running "pytest" interactively. (Of course, the uploaded
artifacts remain useful and are still being generated.)

This also has the benefit of greater accuracy. For example, when
looking for where we install dependencies, someone unfamiliar with
the workflows would not necessarily know to look in "Run Tests".