Add isort config for the import style in use
This project already uses a consistent style for ordering and
formatting imports, but it is not the style isort uses by default.
This configures isort to use that style, fixing it both on CI and
when run manually.

The main configuration change is to how imports are formatted when
enough names are imported `from` a module that they have to be
spanned across multiple lines. This uses the same line length max
we're using with black, and makes it so that when a `from` import
does have to be spanned across multiple lines, it lists one name
per line. That is what we're already doing, and it has the
advantage of producing readable diffs and fewer merge conflicts.

The "atomic" setting for isort is not important on CI, where we're
just checking and not making changes, but it is a safeguard against
commands like "isort ." accidentally introducing syntax errors.