Add dev container configuration
This adds a dev container configuration for Python 3.11. The reason
for this is that the universal image used when creating a codespace
(or local dev container) for a repository that doesn't have a dev
container configuration ships a Python built without lzma. For this
project, that prevents some tests from being able to run, and also
prevents experimentation with lzma compression.