Add CI test jobs for other operating systems
This expands the CI test matrix so that in addition to testing all
three Python versions we support on Ubuntu, it also tests them on
macOS and on Windows. This expands the number of jobs from three to
nine.

The output artifact names are expanded accordingly, so that they do
not clash with each other. A less importand corresponding change is
also made to the artifact name in the publishing workflow for
consistency (just as it already had `3.9` in its name before, for
consistency).

It would be more common and idiomatic to have the `os` matrix
variable take on full names like `ubuntu-latest` rather than
`ubuntu`. However, examining output would be cumbersome if `latest`
appeared in all the artifact filenames. Using the shorter names in
`os` and specifying the OS as `${{ matrix.os }}-latest` avoids that
without extra complexity.

The setup-python action on Windows doesn't (or doesn't always)
supply a `python3` executable, but it supplies a `python`
executable on all OSes (at least for CPython, as we are using). So
this changes the poetry installation command in the test workflow
so it pipes the downloaded script to `python` instead of `python3`.