Change how we install Poetry for portability
This changes how `poetry` is obtained in the test workflow, because
on macOS the Poetry installation script puts it in a directory
where it is not found. It's possible to hard-code or obtain the
location and add it to `$PATH`, but this would make the workflow
more complicated and harder to understand, or be brittle, or both.

Installing it with `pip` in the global environment makes it work on
all OSes without further configuration. It's moderately undesirable
to do this, because generally it's better to avoid putting things
in the global environment. However, because we are using a virtual
environment to install the project and run the tests, sufficient
isolation is still maintained in this case.

The reason this uses pip instead of pipx, even though pipx is
usually the ideal solution to such problems, is that pipx on a
GitHub CI runner defaults to using the system Python (which it was
installed with) even when setup-python has been used to get a
different version. The effect is to install a project and run its
tests in a way that looks good but actually uses the same version
of Python in every generated job, unless one is extremely careful
about how one runs pipx and/or poetry (which itself may increase
workflow complexity).

Although using `pip install poetry` may be justified for the above
reasons, it is not ideal, so I have not made that change in the
publishing workflow. Perhaps a better way that is not excessively
complicated can be found in the future and used in both places.