Wrap original codebase docstrings when >88 columns
Code style in the original codebase, including wrapping function
definitions and calls and regrouping/reordering imports, was
reglarized in bf86554 (as part of #30). But line lengths in
docstrings were not changed (because black does not automate this
and it is unrelated to isort).

This tweaks docstrings in modules in original_codebase/ so they're
at most 89 columns. In one case, I removed a small amount of extra
wording from a docstring's description of function parameters. In
all other cases, I wrapped parts of docstrings that were over 89
columns to 79 columns, while leaving other parts (even in the same
docstrings and between 79 and 89 columns) as they were before. When
wrapping descriptions of function parameters, I used hanging
indents, for readability and because that style is the most
consistent with nearby code (and code outside original_codebase/).