Put upper bound back to 50
This puts the upper bound for the randomly determined sample size
back to 50. The immediately preceding commits verified that the
bug can be fixed by a change in the lower bound, but we don't
always want to use a lower bound of 2!

It may be best to choose the sample size non-randomly, perhaps
testing multiple sample sizes per run using parametrization or
subtests. But that is more extensive than the small change needed
for #45, and thus beyond the scope here.